date/hour read w

+++ I2C.h
@@ -30,7 +30,7 @@ int8_t init_i2c ();
  \param[in] dataSize
  \return TRUE if there was no problem
  */
-int8_t i2c_read (uint8_t slaveAdress, uint16_t subaddress, uint8_t dataSize,
+int8_t i2c_read (uint8_t slaveAdress, uint8_t subaddress, uint8_t dataSize,
         uint8_t* bufferOut, uint8_t subaddressSize);
 /***************************************************************************/
 /***************************************************************************/
@@ -43,7 +43,7 @@ int8_t i2c_read (uint8_t slaveAdress, uint16_t subaddress, uint8_t dataSize,
  \param[in] buffer[dataSize] cointains the information to write
  \return TRUE if there was no problem
  */
-int8_t i2c_writes (uint8_t slaveAdress, uint16_t subaddress, uint8_t dataSize,
+int8_t i2c_writes (uint8_t slaveAdress, uint8_t subaddress, uint8_t dataSize,
         uint8_t* buffer, uint8_t subaddressSize);
 
 #endif /* I2C_H_ */
